+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Expose method `icrc3_get_blocks` in class `IcrcLedgerCanister`.
 - Add `setFollowing` to sns governance canister.
 - New service util `queryAndUpdate` that aids in executing a request and handling the results, for both type of calls.
+- Added type predicate to narrow value to `undefined`, `null`, or empty string in type guards in util `isEmptyString`.
 
 # 2025.03.10-1330Z